+++ sudsoln/__init__.py
@@ -1,8 +1,4 @@
 
 from sudsoln.sudoku import *
-import sudsoln.appearance as appearance
-import sudsoln.candidate as candidate
 import sudsoln.questions as questions
-import sudsoln.sarray as sarray
-import sudsoln.union as union
-__version__ = '0.0.14'
+__version__ = '0.0.15'