+++ sudsoln/tests.py
@@ -1,11 +1,9 @@
 
 import unittest
-import appearance
 import candidate
 import questions
 import sarray
 import sudoku
-import union
 
 
 # START: testing Array ###################################################